Refactored StartTournamentHandler and added new service RockPaperArenaService
